requirements.txt
argparse==1.4.0
docker==7.1.0
PyYAML==6.0.1
mergedeep==1.3.4
GitPython==3.1.43